+++ python/ray/__init__.py
@@ -253,7 +253,6 @@ __all__ += [
     "autoscaler",
     "data",
     "internal",
-    "sandbox",
     "util",
     "widgets",
     "workflow",
@@ -283,14 +282,13 @@ __all__ += [
 if TYPE_CHECKING:
     from ray import autoscaler
     from ray import data
-    from ray import sandbox
     from ray import workflow
 else:
 
     def __getattr__(name: str):
         import importlib
 
-        if name in ["data", "workflow", "autoscaler", "sandbox"]:
+        if name in ["data", "workflow", "autoscaler"]:
             return importlib.import_module("." + name, __name__)
         raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
