--- python/ray/sandbox/_internal/__init__.py
@@ -1,1 +1,0 @@
-# Internal utilities for ray.sandbox
